+++ habana_eager/ops/eager_op.cpp
@@ -36,7 +36,7 @@ void EagerLoweringTask(
       true,
       "EagerLoweringTask()",
       eager_op_meta_data.op_name_,
-      (int32_t)LOP::PipelineStageID::PIPELINE_STAGE_LOWERING_ID,
+      LOP::PipelineStageID::PIPELINE_STAGE_LOWERING_ID,
       lowering_queue_length);
   habana::eager::EagerExec hlexec{
       std::move(symbol),
